add test

+++ pet-clinic-api/src/main/java/com/samuel/sniffers/service/impl/CustomerServiceImpl.java
@@ -28,10 +28,9 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-@Transactional(isolation = Isolation.READ_COMMITTED)
 public class CustomerServiceImpl implements CustomerService {
 
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+    private final Logger logger;
     private final CustomerRepository customerRepository;
     private final SecurityService securityService;
     private final EntityFactory entityFactory;
@@ -41,9 +40,11 @@ public class CustomerServiceImpl implements CustomerService {
         this.customerRepository = customerRepository;
         this.securityService = securityService;
         this.entityFactory = entityFactory;
+        this.logger = LoggerFactory.getLogger(this.getClass());
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public CustomerResponseDTO create(CustomerDTO dto) {
         final String currentCustomerToken = securityService.getCurrentCustomerToken();
 
@@ -61,11 +62,13 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public CustomerResponseDTO findById(String customerId) {
         return entityFactory.convertToDTO(getCustomer(customerId), CustomerResponseDTO.class);
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public List<CustomerResponseDTO> findAll() {
         String token = securityService.getCurrentCustomerToken();
         return customerRepository.findAllWithAccess(token, securityService.isAdmin(token))
@@ -75,6 +78,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public CustomerResponseDTO update(String customerId, CustomerDTO dto) {
         Customer customer = getCustomer(customerId);
         customer.setName(dto.getName());
@@ -84,6 +88,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public CustomerResponseDTO update(String customerId, CustomerPatchDTO dto) {
 
         if (dto.getName() == null && dto.getTimezone() == null) {
@@ -96,6 +101,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public CustomerBatchUpdateResponseDTO batchUpdate(CustomerBatchUpdateDTO dto) {
 
         final List<String> ids = dto.getUpdates().stream()
@@ -158,7 +164,14 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public void delete(String customerId) {
+
+        if (!customerExist(customerId)) {
+            logger.error("Customer with id: {} not found", customerId);
+            throw new ResourceNotFoundException("Customer not found");
+        }
+
         customerRepository.delete(getCustomer(customerId));
         logger.error("Customer with id: {} deleted successfully", customerId);
     }
